accounts/serializers.py
from rest_framework import serializers
from .models import CustomUser, WorkExperience, Education, Skill
from django.contrib.auth import authenticate
from rest_framework.exceptions import AuthenticationFailed
from rest_framework_simplejwt.tokens import RefreshToken



class RegisterSerializer(serializers.ModelSerializer):
    confirm_password = serializers.CharField(write_only=True)

    class Meta:
        model = CustomUser
        fields = [
            'first_name', 'last_name', 'email', 'phone_number',
            'password', 'confirm_password', 'agreed_to_terms'
        ]
        extra_kwargs = {'password': {'write_only': True}}

    def validate(self, data):
        if data['password'] != data['confirm_password']:
            raise serializers.ValidationError("Passwords do not match")
        if not data.get('agreed_to_terms'):
            raise serializers.ValidationError("You must agree to the terms")
        return data

    def create(self, validated_data):
        validated_data.pop('confirm_password')  # ✅ Remove non-model field
        validated_data['username'] = validated_data.get('email')  # ✅ Required for create_user
        user = CustomUser.objects.create_user(**validated_data)
        return user

class LoginSerializer(serializers.Serializer):
    email = serializers.EmailField()
    password = serializers.CharField(write_only=True)
    
    def validate(self, data):
        user = authenticate(email=data['email'], password=data['password'])
        if not user:
            raise AuthenticationFailed("Invalid email or password")
        if not user.is_active:
            raise AuthenticationFailed("Account is disabled")
        
        refresh = RefreshToken.for_user(user)
        return {
            'access': str(refresh.access_token),
            'refresh': str(refresh),
            'user': {
                'id': user.id,
                'email': user.email,
                'first_name': user.first_name,
                'last_name': user.last_name
            }
        }


class ProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = [
            'first_name', 'last_name', 'email', 'phone_number',
            'profile_score', 'applications_count', 'interviews_count',
            'resume_file', 'profile_picture'
        ]


class WorkExperienceSerializer(serializers.ModelSerializer):
    class Meta:
        model = WorkExperience
        fields = [
            'id', 'title', 'company', 'location', 'start_date', 'end_date',
            'is_current', 'description', 'created_at', 'updated_at'
        ]
        read_only_fields = ['created_at', 'updated_at']


class EducationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Education
        fields = [
            'id', 'degree', 'school', 'start_year', 'end_year',
            'is_current', 'grade', 'description', 'created_at', 'updated_at'
        ]
        read_only_fields = ['created_at', 'updated_at']


class SkillSerializer(serializers.ModelSerializer):
    level_display = serializers.CharField(source='get_level_display', read_only=True)
    level_percentage = serializers.IntegerField(read_only=True)
    
    class Meta:
        model = Skill
        fields = [
            'id', 'name', 'level', 'level_display', 'level_percentage',
            'created_at', 'updated_at'
        ]
        read_only_fields = ['created_at', 'updated_at']
    
    def validate_name(self, value):
        """Ensure skill name is not empty and properly formatted"""
        value = value.strip()
        if not value:
            raise serializers.ValidationError("Skill name cannot be empty")
        return value.title()  # Capitalize first letter of each word
