requirements_industry.txt
# SkillConnect - Industry Level Requirements
# Current Requirements + Professional Additions

# Current Backend Requirements
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
Pillow==10.0.1
djangorestframework-simplejwt==5.3.0

# Industry Level Additions

# API Documentation
drf-yasg==1.21.7                    # Swagger/OpenAPI documentation
django-extensions==3.2.3            # Enhanced management commands

# Performance & Caching
django-redis==5.4.0                 # Redis caching backend
redis==5.0.1                        # Redis client
django-debug-toolbar==4.2.0         # Performance debugging

# Testing & Quality
coverage==7.3.2                     # Code coverage reports
factory-boy==3.3.0                  # Test data factories
pytest-django==4.6.0                # pytest integration
flake8==6.1.0                       # Code linting
black==23.9.1                       # Code formatting

# Security Enhancements
django-ratelimit==4.1.0             # API rate limiting
django-guardian==2.4.0              # Object-level permissions
cryptography==41.0.7                # Advanced encryption

# Monitoring & Logging
sentry-sdk==1.38.0                  # Error tracking
django-prometheus==2.3.1            # Metrics collection

# Email & Notifications
django-anymail==10.2                # Email service integration
celery==5.3.4                       # Background task processing
django-celery-beat==2.5.0           # Periodic tasks

# File Storage
django-storages==1.14.2             # Cloud storage backends
boto3==1.34.0                       # AWS S3 integration

# Real-time Features
channels==4.0.0                     # WebSocket support
channels-redis==4.1.0               # Redis channel layer

# Third-party Integrations
requests==2.31.0                    # HTTP requests
social-auth-app-django==5.4.0       # OAuth integration
python-decouple==3.8                # Environment variables

# AI/ML Integration
openai==1.3.7                       # OpenAI GPT API
pandas==2.1.3                       # Data analysis
scikit-learn==1.3.2                 # Machine learning

# Production Deployment
gunicorn==21.2.0                    # WSGI server
whitenoise==6.6.0                   # Static file serving
psycopg2-binary==2.9.9              # PostgreSQL adapter